+++ lib/OrdinaryDiffEqMultirateImplicit/Project.toml
@@ -1,33 +1,47 @@
-name = "OrdinaryDiffEqMultirate"
-uuid = "d4b830b4-ac80-426b-8507-16693d424963"
+name = "OrdinaryDiffEqMultirateImplicit"
+uuid = "fd3f87d1-5955-4d40-9638-6d84dd605b32"
 authors = ["singhharsh1708 <hs1663531@gmail.com>"]
-version = "3.0.0"
+version = "1.0.0"
 
 [deps]
+ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 MuladdMacro = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
 OrdinaryDiffEqCore = "bbf590c4-e513-4bbe-9b18-05decba2e5d8"
+OrdinaryDiffEqDifferentiation = "4302a76b-040a-498a-8c04-15b101fed76b"
+OrdinaryDiffEqMultirate = "d4b830b4-ac80-426b-8507-16693d424963"
+OrdinaryDiffEqNonlinearSolve = "127b3ac7-2247-4354-8eb6-78cf4e7c58e8"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
+SciMLOperators = "c0aeaf25-5076-4817-a8d5-81caf7dfa961"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 SciMLTesting = "2.1"
+ADTypes = "1.22.0"
 Aqua = "0.8.11"
 DiffEqBase = "7"
 DiffEqDevTools = "3"
 FastBroadcast = "1.3"
 JET = "0.9, 0.11"
+LinearSolve = "5.1"
 MuladdMacro = "0.2.4"
-OrdinaryDiffEqCore = "4"
+OrdinaryDiffEqCore = "4.6"
+OrdinaryDiffEqDifferentiation = "3.5"
 OrdinaryDiffEqLowOrderRK = "2"
+OrdinaryDiffEqMultirate = "3"
+OrdinaryDiffEqNonlinearSolve = "2.5"
 Pkg = "1"
 RecursiveArrayTools = "4.2.0"
 Reexport = "1.2.2"
 SafeTestsets = "0.1.0"
 SciMLBase = "3.39"
+SciMLOperators = "1.24.4"
+SparseArrays = "1.10"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
@@ -51,8 +65,17 @@ path = "../DiffEqDevTools"
 [sources.OrdinaryDiffEqCore]
 path = "../OrdinaryDiffEqCore"
 
+[sources.OrdinaryDiffEqDifferentiation]
+path = "../OrdinaryDiffEqDifferentiation"
+
 [sources.OrdinaryDiffEqLowOrderRK]
 path = "../OrdinaryDiffEqLowOrderRK"
+
+[sources.OrdinaryDiffEqMultirate]
+path = "../OrdinaryDiffEqMultirate"
+
+[sources.OrdinaryDiffEqNonlinearSolve]
+path = "../OrdinaryDiffEqNonlinearSolve"
 
 [targets]
 test = ["DiffEqDevTools", "LinearAlgebra", "OrdinaryDiffEqLowOrderRK", "SafeTestsets", "Test", "Pkg", "SciMLTesting"]
