+++ lib/OrdinaryDiffEqMultirateImplicit/test/qa/Project.toml
@@ -4,6 +4,7 @@ DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 OrdinaryDiffEqCore = "bbf590c4-e513-4bbe-9b18-05decba2e5d8"
 OrdinaryDiffEqMultirate = "d4b830b4-ac80-426b-8507-16693d424963"
+OrdinaryDiffEqMultirateImplicit = "fd3f87d1-5955-4d40-9638-6d84dd605b32"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
@@ -11,8 +12,12 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 AllocCheck = "0.2"
 OrdinaryDiffEqCore = "4"
 OrdinaryDiffEqMultirate = "3"
+OrdinaryDiffEqMultirateImplicit = "1"
 SciMLBase = "3"
 julia = "1.10"
 
 [sources.OrdinaryDiffEqCore]
 path = "../../../OrdinaryDiffEqCore"
+
+[sources.OrdinaryDiffEqMultirate]
+path = "../../../OrdinaryDiffEqMultirate"
